Enhance event and operation lifecycle with new fields and refactorings
- Added message and level fields to the snapshot structure for improved event context.
- Updated snapshot method to include new fields for better event handling.
- Refactored Commit function to utilize the new snapshot structure for message retrieval.
- Introduced new operation lifecycle management files to support non-HTTP flows and improve operation handling.
